Update split pkg API
